+++ src/main/java/com/example/leaflog/bc/account/auth/infrastructure/security/oauth/parse/GithubOauthUserInfoParser.java
@@ -1,6 +1,7 @@
 package com.example.leaflog.bc.account.auth.infrastructure.security.oauth.parse;
 
 
+import com.example.leaflog.bc.account.auth.application.port.out.GithubEmailPort;
 import com.example.leaflog.bc.account.auth.infrastructure.security.oauth.dto.OauthUserInfo;
 import lombok.RequiredArgsConstructor;
 
@@ -9,7 +10,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class GithubOauthUserInfoParser implements OauthUserInfoParser {
 
-    private final GithubEmailFetcher emailFetcher;
+    private final GithubEmailPort githubEmailPort;
 
     @Override
     public OauthUserInfo parse(Map<String, Object> attributes, String accessToken) {
@@ -18,7 +19,7 @@ public class GithubOauthUserInfoParser implements OauthUserInfoParser {
                 ? (String) attributes.get("name")
                 : (String) attributes.get("login");
 
-        String email = emailFetcher.fetchPrimaryEmail(accessToken);
+        String email = githubEmailPort.fetchPrimaryEmail(accessToken);
 
         String avatarUrl = (String) attributes.get("avatar_url");
