+++ src/main/java/com/example/leaflog/bc/account/auth/presentation/adapter/in/oauth/handler/OauthSuccessHandler.java
@@ -1,8 +1,8 @@
-package com.example.leaflog.bc.account.auth.infrastructure.security.oauth.handler;
+package com.example.leaflog.bc.account.auth.presentation.adapter.in.oauth.handler;
 
+import com.example.leaflog.bc.account.auth.application.port.in.TokenGenerateUseCase;
 import com.example.leaflog.bc.account.auth.infrastructure.security.auth.AuthDetails;
-import com.example.leaflog.bc.account.auth.infrastructure.security.jwt.JwtTokenProvider;
-import com.example.leaflog.bc.account.auth.application.dto.TokenResponseDto;
+import com.example.leaflog.bc.account.auth.application.dto.TokenResponse;
 import com.example.leaflog.bc.account.user.domain.User;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -19,7 +19,7 @@ import java.nio.charset.StandardCharsets;
 @Component
 @RequiredArgsConstructor
 public class OauthSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
-    private final JwtTokenProvider jwtTokenProvider;
+    private final TokenGenerateUseCase tokenGenerateUseCase;
     private final ObjectMapper objectMapper;
 
     @Override
@@ -31,10 +31,7 @@ public class OauthSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
         AuthDetails authDetails = (AuthDetails) authentication.getPrincipal();
         User user = authDetails.getUser();
 
-        TokenResponseDto tokenResponse = TokenResponseDto.of(
-                jwtTokenProvider.generateAccessToken(user.getPrincipalEmail()),
-                jwtTokenProvider.generateRefreshToken(user.getPrincipalEmail())
-        );
+        TokenResponse tokenResponse = tokenGenerateUseCase.generateToken(user.getPrincipalEmail());
 
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.setCharacterEncoding(StandardCharsets.UTF_8.name());
